solaris: fix some printf warnings

+++ hwloc/topology-solaris.c
@@ -718,7 +718,7 @@ hwloc_look_kstat(struct hwloc_topology *topology)
 	    break;
 #endif
 	  default:
-	    fprintf(stderr, "chip_id type %d unknown\n", stat->data_type);
+	    fprintf(stderr, "chip_id type %u unknown\n", (unsigned) stat->data_type);
 	    look_chips = 0;
 	    continue;
 	}
@@ -768,7 +768,7 @@ hwloc_look_kstat(struct hwloc_topology *topology)
 	    break;
 #endif
 	  default:
-	    fprintf(stderr, "core_id type %d unknown\n", stat->data_type);
+	    fprintf(stderr, "core_id type %u unknown\n", (unsigned) stat->data_type);
 	    look_cores = 0;
 	    continue;
 	}
